使用flask-login进行后台管理

+++ initdb.py
@@ -3,13 +3,23 @@ crate test database
 '''
 import os
 from datetime import datetime, timedelta
+from werkzeug.security import generate_password_hash
+import getpass
 from official import app
 from config import db
-from official.views import Stb
+from official.views import Stb, User
 
 def build_sample_db():
+    name = raw_input('user name:')
+    email = raw_input('enter email:')
+    password = getpass.getpass('enter password:')
+    tmp = getpass.getpass('enter password again:')
+    if password!=tmp:
+        print('password not equal')
+        return
     db.drop_all()
     db.create_all()
+    user = User(name, email, generate_password_hash(password))
     stb1 = Stb('T2', 'http://img.tvhome.com/uploads/2014/09/10/19290aff2d718cbed9265f0266b64bc7.jpg',
                260.0, 'http://www.taixin.cn', 'taixin', 'DVB+OTT', datetime.utcnow()+timedelta(hours=1))
     stb2 = Stb('X1', 'http://www.taixin.cn/uploadfiles/image/20141226/20141226132514_6076.jpg',
@@ -19,6 +29,7 @@ def build_sample_db():
     db.session.add(stb1)
     db.session.add(stb2)
     db.session.add(stb3)
+    db.session.add(user)
     db.session.commit()
     return
